[show-history-by-list] Show RecyclerView(WIP because the list is not shown)
